Add support for NodeCondition logic junctions
Within a data model, a node existence condition was limited to one logic clause
(i.e., a Node on which a NodeCondition() has to be verified).

This commit add support for chaining logic clauses with an AND operator or with
an OR operator. To benefit from these new features, the data model keywords
'exists_if/and' and 'exists_if/or' have been defined. They expect a list of logic clauses.