Clarify different types of timeout + Add CallBackOps.Set_FbkTimeout
When a target is loaded, the feedback gathering time duration (fbk timeout) is
retrieved from the target itself (in the project file). It is now possible to change it
dynamically thanks to FmkPlumbing.set_feedback_timeout().

Additionally Data() optional callbacks can now set the feedback timeout that would be
taken into account for further data sending.

Note also that the health-check timeout is calculated from this feedback timeout