Make the Scenario Infrastructure state-machine oriented
A scenario is now a state-machine backed by a graph of steps where
the transitions are guarded by the callbacks previously defined at the step level.

To construct a graph of step, the Step class has a new method .connect_to()
that enables to connect to another step. If callbacks are specified, they will be
checked to verify if the transition can be crossed (False --> don't cross). If there is no callback,
the transition is considered to not be guarded (thus, crossing it is authorized).

Only one transition is chosen at each step. It is the first one, by order of registration,
that can be activated (at least one callback that returns True, or no callback at all).