Revamp String encoding to fix inconsistencies and clumsiness
Python codecs ('latin-1', 'utf8', ...) are now handled directly by the String class
through the parameter 'codec'.
Besides, as this handling is independent from Encoder()s (e.g., GZIP, ...),
the Encoder() associated to an *Encoded* String() apply to the content of
the String(), which is encoded with the python codec provided as parameter.