Update/Fix counter-intuitive Node.set_values()/... behavior
A Node on which one of the following methods were called:
- set_subnodes_basic()
- set_subnodes_with_csts()
- set_values()
- set_generator_func()
- set_func()
lost all their NodeInternals attributes, like their possible constraints with the other
nodes in the graph.
This method take a new parameter 'preserve_node' (defaults to True) which enables
to preserve previous non-specific NodeInternals attributes when these methods are called.
This enable simpler modification of a graph.

Note on disruptors:
- tTYPE leverages that behavior change.
- SIZE and C will now avoid to break a graph by calling such methods, which enables the user
  to call the disruptor FIX after them and obtains the desired result.