Fix PPPoE DM scenario
FIX_FIELDS disruptor should not unfreeze the 'length' as it is useless since 0.25 because
the generator MH.LEN is non-freezable by default.

Besides, this triggers a bad behavior with tTYPE (since 0.25) as it leads to the creation
of a new node (by resetting the generator --- custo.ResetOnUnfreeze is True by default)
that tTYPE is not aware of. Thus, seeing a new node at the next step, tTYPE decides
to fuzz it, which leads to an endless loop.