Make the disruptor C more respectful of the data model
In the previous implementation the C disruptor changed the selected nodes
by String-typed ones (with the computed corrupted value) with no regard
for the previous type. This led to data model inconsistencies,
especially when some nodes depended on the type of the corrupted one.

Fix it by making C leveraging the absorption operation.