Make send_loop.aligned disabled by default
The issue fixed by the commit b7e28d319acde860c46bdcd32ad2308b8229d8af
is able to produce with python3 a traceback that is not properly
displayed on screen.

Hence, the buffered output will be disabled by default until all
tracebacks are properly displayed on screen, in order to facilitate
debugging when we can not afford to reproduce the error by trying to
trigger the bug without the buffering enabled.