Add support to use multiple targets at the same time
Implement support to enable multiple targets at the same time, meaning
it is now possible to stimulate multiple targets from the same instance
of fuddly.
It makes possible to sequence different actions through various systems
or on the same system but through different kinds of interfaces
(represented by different targets); and to react on feedback retrieved
from all the stimulated targets in a centralized way.

- The interaction with the logger/fmkDB has been updated to support multi-targets
- All Fmkshell target-related commands have been updated to leverage this feature
  (run_project, reload_all, send*, load_targets, ...).
- The Scenario infrastructure is updated to leverage multi-targets. Steps
  can stimulate specific targets through usage of virtual target IDs.
  Then a mapping can be configured with real target in project files
  that wants to use the scenario.
- The Operator support this feature with a new parameter in Operation class
  constructor.