Scenario: Fix step.content inconsistent side-effect
When a callback modify the object referenced by step.content, but does
not assign it back to step.content, the change are lost when the data
is retrieved by the framework, because the framework use step._data_desc
and this variable does not refer to the modified object.
To avoid this inconsistency, the property step.content modify
step._data_desc before returning.