Scenario: Add new condition to cross a transition
For step hosting a DataProcess, a transition can now be guarded by a specific
condition that is satisfied when no more data can be generated from the
DataProcess. It has to be specified while calling Step.connect_to() method
by setting the parameter dp_completed_guard to True.

Scenario documentation is updated.