[API Change] Node dictionary interface update
When a string is provided as a key to Node.__getitem__(), it is interpreted
as a path regexp in the node-based data representation.

Previously the first Node matching the path was returned or None if
there was not a match. Now, a list will be returned providing all the nodes
and None will be return when nothing match.

For the other key types, a list was already returned. But also in the case
where nothing match. Thus, for consistency, None is now returned
in the latter case.

Every users of this interface have been updated as well as the documentation/