[API change] node searching primitives new default behavior + other fixes/enhancements
* Node searching primitives could have side effects, namely freezing part of the node graph in the process.
This was caused by the resolution of Generator-nodes taking nodes parameter (if present in the graph), as
in the process of resolving/expanding the Generator-nodes the parameters are also resolved thus frozen.
--> To have control on this side effect a new param has been added to the node searching primitives: resolve_generator.
It defaults to False in order to avoid side effect, but it will also affect searches as nothing will be
retrieved by default from the generator nodes.
WARNING: the default behavior of searching primitives is thus changed. Node.__getitem__() is also defaulting to the new behavior.
To use the old behavior, set resolve_generator to True.

* JSON data model enhancements and fix
* Improve absorption of INT_str

* NetworkTarget new parameter 'listen_on_start': If `True`, servers will be launched right after the NetworkTarget
starts. Otherwise, they will be launched in a lazy mode, meaning just when something is
about to be sent through the server mode interface.