Fix/improve default behavior of node searching primitives
Default behavior of these primitives avoid side effects. This is why
@resolve_generator parameter is set to False by default.
However, this parameter when set to False was restricting the searching too much.
It made searching primitives ignoring generator nodes at all.
Now, if resolve_generator is `False`, generator nodes won't be resolved, but if they already are in a
resolved state before a searching primitive is called on them, it will work as if @resolve_generator
is set to True. Indeed no side effects could result from the call of a searching primitve in this case.