Bugfix: DataProcess handling
When a DataProcess contains multiple processes, all the processes
are walked through in order to complete the DataProcess.
Before this fix, in some situation, we could miss the completion indicator
and walk again to the list of processes (never ending the DP),
as if @auto_regen was set to True.