Node: fix incorrect handling of separator in some situation
When the 'separator' keyword was used in a non-terminal node with the suffix parameter set to False,
and when the last child of the NT node had an existence condition that was evaluated to False,
then the NT node resolved incorrectly with a separator at the end.

Fix this situation by performing suffix handling at the end of NT resolution.