Improve tWALK exploration capability
When a data model uses some existence conditions, walking through the graph can be tricky.
As a result, some data compliant with the data model could never be generated by tWALK.
The walking algorithm has been improved in order to take into account the changes that
may occur among the sibbling nodes of the node which is currently walking through.
And in the case there is some change, then the algorithm will walk through the new nodes.
This improves the exploration capability of tWALK.
A new parameter to tWALK, @consider_sibbling_change (default to True), has been introduced
in order to stick to the old behavior if needed.