Fix a usability issue when CSP is used in a DM
The Generators which are automatically created for each atom registered in a DM
have their `freeze` parameter set to False by default. As it makes sense when you use
these Generators in conjunction with other operators such as tWALK, it makes surprising situation
when you use these Generators alone and that these Generators includes a CSP (as the
'resolve_csp' parameter will be ignored basically).

So when a Generator is used alone we now always freeze its outcomes.