Add new Test.OnlyCornerCases attribute in knowledge infrastructure
This new information item enables a new mode when fuzzing cases are generated by the
typed nodes within a data model. Only valid corner cases will be generated if this item is
part of the project knowledge base.
(TODO: make all types compliant with this, as only INT_str are currently supported)