Fix issue with constraints backend and tTYPE
An inconsistency may happen when a disruptor play with the domain of a node
which is normally a vt.INT and replace it temporarily for the generated
fuzzing cases by a vt.String.
Make sure to update the z3 variable types when detecting such inconsistencies,
and ignore any defined z3 formulas if they do not make sense anymore.