DM: Fix irrelevant use of color escape sequences within generator nodes
While printing to the console in highlighted mode, some generator nodes can get incorrect values
when they freeze node parameters.
We now disable color mode while computing the generator node, in order to have consistency between
what is printed on the console and what is sent or logged in files.