Fix Existence Condition old bug + improve/fix recursive node generation
* NodeInternals_NonTerm._cleanup_delayed_nodes() could misbehave in some condition.
The parameter @idx was used why it shall have not. Indeed,
if multiple delayed nodes in the same node_list need to be
cleaned up we cannot make assumption on the calling order,
and the idx remains only valid in descending order.

* The misbehavior of _cleanup_delayed_nodes() was made visible through Recursive Node
test cases. After fixing the bug, further integration of Recursive Node was possible.