rework the magic imports
The importer hook will now only ever return namespace specs.
Unless the entry_point is pointing to a module.
It is also now possible to have a virtual submodule for entry_points
making them align with the organisation of data in the user_data_folder
and making imports homogeneous between every module source.