[API change] Change of FmkPlumbing.process_data() return value
process_data() was only returning a Data() or None if ever a data maker yielded or crashed.
In the case a data maker yielded (case of a stateful operator needing a new input to
generate new outputs), process_data() was not explicit if either the current action list provided
in parameter could trigger new outputs, because all operators may not be exhausted.
The user had to execute again process_data() to see what happened.

Such cases happened when at least 2 stateful operators were in the provided action list.
For example: SEED tWALK(path='path/to/a/node/to/walk/into') tTYPE

Now, process_data() is returning a tuple with the Data|None for the first item and
for the second item, a boolean value notifying if the action list is exhausted or not.

This change affects the scenario infrastructure and process_data_and_send() that will exhaust
entirely an action list before returning.

As a consequence, the send* command of the fuddly shell will exhaust entirely the action
list before returning, and the feedback messages provided to the user will inform about that.