Fix loop case in ModelWalker infra with consider_sibbling_change=True
When (in second step) a subnode ask for a reset and it is NOT exhausted,
meaning the reset will trigger a new value/shape, then we now avoid
to unfreeze it through its parent node in the subsequent consider_side_effects_on_sibbling clause.
Indeed, doing so will result in going back to the initial state of the subnode, thus cancelling the
previous action.