Change behavior of customization MutableClone for NT nodes
When the customization MutableClone is set on a non-terminal node, every subnodes with
a qty > 1 will be cloned and then the Mutable attr will be removed so that fuzzing operators
could ignore them (heuristic to minimize the number of cases). But when the customization
CycleClone is also set, this is often irrelevant as the clone will be different
(especially when CycleClone is used to have different node shapes).