Add the new Node attribute "Optional" and make tTYPE leverages it
When a node is defined with a minimum quantity equal to 0, it means it is optional.
When an operator, like tTYPE, is requested to generate invalid data (@only_invalid_cases)
from a String-based node, one case is to replace the current string value with an
empty-string, but in the case the node is optional it is actually not invalid.
Now that this new Optional attribute is visible by any operators, tTYPE is leveraging
it in order to remove the empty-string case when the node is optional.