Add new Node separator attribute @optional_tail + various fixes linked to separator absorption
Used optionally in a separator descriptor within a non-terminal node descriptor. Relevant when @always is True.
If set to True then when subnodes at the tail are not generated, separators are considered to be optional.
Therefore separators could be omitted at the tail when the non-terminal node will be resolved.
This is enforced when NonTermCusto.SepDelOptionalTail is set.